requirements.txt
matplotlib
python-dotenv
requests
google-auth>=2.34.0
numpy
pandas